// src/cell/types.ts
export interface CellColor {
  topLeftHorizontal?: boolean;
  topRightHorizontal?: boolean;
  middleLeftHorizontal?: boolean;
  middleRightHorizontal?: boolean;
  bottomLeftHorizontal?: boolean;
  bottomRightHorizontal?: boolean;
  topLeftVertical?: boolean;
  topRightVertical?: boolean;
  topMiddleVertical?: boolean;
  bottomLeftVertical?: boolean;
  bottomRightVertical?: boolean;
  bottomMiddleVertical?: boolean;
  topLeftDiagonal?: boolean;
  topRightDiagonal?: boolean;
  bottomLeftDiagonal?: boolean;
  bottomRightDiagonal?: boolean;
}